Use TypeScript
Add a TypeScript build that includes type definitions in the package for
upstream TypeScript projects to use. Unknown types are defined as `any`
for now, which we'll fix incrementally.

The only visible change to package callers is the default export has
moved to a named export due to the conflict between standard modules and
CommonJS modules.

    // Before
    const PSDB = require('planetscale-node')
    // After
    const {PSDB} = require('planetscale-node')